Conflicts:
	src/main/java/project/web/mvc/repository/WishListRepository.java
	src/main/webapp/WEB-INF/views/onlecture/list.jsp

	wishlistRepo렁  list.jsp 수정 commit 완료

+++ src/main/java/project/web/mvc/service/WishListService.java
@@ -2,6 +2,6 @@ package project.web.mvc.service;
 
 public interface WishListService {
 
-	public void wishlistInsert(int studentNo, int onLectureNo);
+	public void wishlistInsert(Long studentNo, Long onLectureNo);
 	
 }
